WIP: Integrate buffer
